requirements.txt
# Data handling
pandas>=2.1
numpy>=1.26
pyarrow>=14.0
openpyxl>=3.1
xlrd>=2.0
pyyaml>=6.0
requests>=2.31

# Analysis
scipy>=1.11
statsmodels>=0.14
pmdarima>=2.0

# Modeling
xgboost>=2.0
scikit-learn>=1.3
shap>=0.43

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Progress and logging
tqdm>=4.66

# Development
jupyter>=1.0
ipykernel>=6.0

# Config
python-dotenv>=1.0
